biljka-back/requirements.txt
alembic==1.16.1
annotated-types==0.7.0
anyio==4.9.0
arabic-reshaper==3.0.0
asn1crypto==1.5.1
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cryptography==45.0.3
cssselect2==0.8.0
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.3.0
fastapi==0.115.12
greenlet==3.2.3
h11==0.16.0
html5lib==1.1
idna==3.10
lxml==5.4.0
Mako==1.3.10
MarkupSafe==3.0.2
oscrypto==1.3.0
pillow==11.2.1
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
pyHanko==0.29.0
pyhanko-certvalidator==0.27.0
pypdf==5.6.0
python-bidi==0.6.6
python-dotenv==1.1.0
PyYAML==6.0.2
reportlab==4.4.1
requests==2.32.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
svglib==1.5.1
tinycss2==1.4.0
tomli==2.2.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.4.0
uvicorn==0.34.3
webencodings==0.5.1
xhtml2pdf==0.2.17
